Added: Login validation with firebase, user sesssion, users and permisions, delete timer functionality

+++ config/routes.php
@@ -14,9 +14,11 @@ return function (App $app) {
     $app->group(
         '/api',
         function (\Slim\Routing\RouteCollectorProxy $group) use ($c) {
-            $group->get('/timers', \App\Controller\TimerController::class . ':indexAction');
-            $group->post('/timers', \App\Controller\TimerController::class . ':createAction');
-            $group->put('/timers/{id}', \App\Controller\TimerController::class . ':updateAction');
-            $group->delete('/timers/{id}', \App\Controller\TimerController::class . ':deleteAction');
+            $group->get('/timers', \App\Controller\TimerController::class . ':indexAction')->add(\App\Middleware\PermissionsMiddleware::class);
+            $group->post('/timers', \App\Controller\TimerController::class . ':createAction')->add(\App\Middleware\PermissionsMiddleware::class);
+            $group->put('/timers/{id}', \App\Controller\TimerController::class . ':updateAction')->add(\App\Middleware\PermissionsMiddleware::class);
+            $group->delete('/timers/{id}', \App\Controller\TimerController::class . ':deleteAction')->add(\App\Middleware\PermissionsMiddleware::class);
+            $group->post('/auth/login', \App\Controller\AuthController::class . ':loginAction');
+            $group->get('/auth/logout', \App\Controller\AuthController::class . ':logoutAction');
         });
 };
